Moved Chunk to chunk module
