+++ src/client/input/mod.rs
@@ -25,7 +25,8 @@ use crate::block::{
 use crate::sim::chunk::{
     ChunkPos
 };
-use self::chunk::{Chunk, ChunkData, ChunkState};
+use self::chunk::{ChunkData, ChunkState};
+use crate::sim::chunk::Chunk;
 use crate::config::{load_config, Config};
 use crate::core::messages::client::{ToInput, ToMeshing, ToNetwork};
 use crate::input::KeyboardState;
