+++ src/sim/chunk/mod.rs
@@ -1,5 +1,6 @@
-use crate::block::BlockId;
+use crate::block::{Block, BlockId, BlockRegistry};
 use crate::CHUNK_SIZE;
+use crate::Vertex;
 use serde_derive::{Deserialize, Serialize};
 use hashbrown::hash_map::HashMap;
 use hashbrown::hash_map::Entry;
@@ -8,6 +9,7 @@ use hashbrown::hash_map::DefaultHashBuilder;
 pub type BlockData = BlockId;
 pub type ChunkFragment = [BlockData; CHUNK_SIZE];
 pub type ChunkArray = [[ChunkFragment; CHUNK_SIZE]; CHUNK_SIZE];
+pub type ChunkSidesArray = [[[u8; CHUNK_SIZE]; CHUNK_SIZE]; CHUNK_SIZE];
 
 mod pos;
 pub use pos::*;
@@ -141,4 +143,41 @@ impl ChunkState {
         *self = ChunkState::Modified(arr.clone(), v + 1);
     }
 
+}
+
+/// Chunk type
+#[derive(Clone, Debug)]
+pub struct Chunk {
+    /// Blocks in the chunk
+    pub blocks: Box<ChunkArray>,
+    /// Empty blocks adjacent to the chunk (1 is for non-opaque, 0 is for opaque)
+    pub sides: Box<ChunkSidesArray>,
+}
+
+impl Chunk {
+    pub fn new() -> Chunk {
+        Chunk {
+            blocks: Box::new([[[BlockId(0); CHUNK_SIZE]; CHUNK_SIZE]; CHUNK_SIZE]),
+            sides: Box::new([[[0b00000000; CHUNK_SIZE]; CHUNK_SIZE]; CHUNK_SIZE]),
+        }
+    }
+
+    pub fn calculate_mesh(&self, blocks: &BlockRegistry) -> Vec<Vertex> {
+        let mut vec: Vec<Vertex> = Vec::new();
+        for i in 0..CHUNK_SIZE {
+            for j in 0..CHUNK_SIZE {
+                for k in 0..CHUNK_SIZE {
+                    // Don't render hidden blocks
+                    if self.sides[i][j][k] != 0xFF {
+                        blocks.get_block(self.blocks[i][j][k]).render(
+                            &mut vec,
+                            self.sides[i][j][k],
+                            [i as u64, j as u64, k as u64],
+                        );
+                    }
+                }
+            }
+        }
+        vec
+    }
 }
